+++ crypto-sodium/CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 * `sodiumInit`
+* Random bytes generation: `Crypto.Random`